+++ tests/Mocks/SampleIntBackedEnum.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Plank\Metable\Tests\Mocks;
+
+enum SampleIntBackedEnum: int
+{
+    case One = 1;
+}