+++ tests/Mocks/SamplePureEnum.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Plank\Metable\Tests\Mocks;
+
+enum SamplePureEnum
+{
+    case Alpha;
+}