+++ tests/Mocks/SampleStringBackedEnum.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Plank\Metable\Tests\Mocks;
+
+enum SampleStringBackedEnum: string
+{
+    case Alpha = 'alpha';
+    case Numeric = '1';
+
+}